// src/csv/interfaces/store-csv-row.interface.ts
export interface StoreCsvRow {
    id_almacen: string;
    nombre_almacen: string;
    direccion: string;
    ciudad: string;
    departamento: string;
    pais: string;
    codigo_postal: string;
    latitud: string;
    longitud: string;
    gerente: string;
    telefono: string;
    email: string;
    capacidad_m2: string;
    estado: string;
}